# openapi_with_crud.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Microservicios Aritméticos + CRUD/Logs – OpenAPI Unificado",
    "version": "1.1.1",
    "description": "12 microservicios (suma, resta, multiplicación, división) en puertos 3000–3011, más CRUD/Logs en puerto 4010."
  },
...
  "components": {
    "schemas": {
...
      "LogEntry": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "op": { "type": "string", "enum": ["SUMA","RESTA","MULT","DIV","CRUD"] },
          "num1": { "type": "number", "nullable": true },
          "num2": { "type": "number", "nullable": true },
          "result": { "type": "number", "nullable": true },
          "source": { "type": "string", "enum": ["query","path","body","crud"] },
          "endpoint": { "type": "string" },
          "method": { "type": "string" },
          "status_code": { "type": "integer" },
          "client_ip": { "type": "string" },
          "payload_json": { "type": "object", "additionalProperties": true, "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "LogCreate": {
        "type": "object",
        "properties": {
          "op": { "type": "string", "enum": ["SUMA","RESTA","MULT","DIV","CRUD"], "default": "CRUD" },
          "num1": { "type": "number", "nullable": true },
          "num2": { "type": "number", "nullable": true },
          "result": { "type": "number", "nullable": true },
          "source": { "type": "string", "enum": ["query","path","body","crud"], "default": "crud" },
          "endpoint": { "type": "string", "example": "/records" },
          "method": { "type": "string", "example": "POST" },
          "status_code": { "type": "integer", "example": 201 },
          "payload": { "type": "object", "additionalProperties": true, "nullable": true }
        }
      },
      "OkLog": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean", "example": true },
          "message": { "type": "string", "example": "Log registrado" }
        }
      },
      "PaginatedLogs": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean", "example": true },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/LogEntry" } },
          "meta": {
            "type": "object",
            "properties": {
              "page": { "type": "integer", "example": 1 },
              "limit": { "type": "integer", "example": 20 }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/ops/log": {
      "get": {
        "summary": "Consultar logs de operaciones",
        "tags": ["logs"],
        "servers": [{ "url": "http://127.0.0.1:4010" }],
        "parameters": [
          { "name": "op", "in": "query", "schema": { "type": "string", "enum": ["SUMA","RESTA","MULT","DIV","CRUD"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 20 } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaginatedLogs" } } } }
        }
      },
      "post": {
        "summary": "Registrar log de operación",
        "tags": ["logs"],
        "servers": [{ "url": "http://127.0.0.1:4010" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LogCreate" } } } },
        "responses": {
          "201": { "description": "Registrado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OkLog" } } } }
        }
      }
    }
  }
}